once_event_test.go
package gosync

import (
	"testing"
)

func TestOnceEvent_HasFired(t *testing.T) {
	e := NewOnceEvent()
	if e.HasFired() {
		t.Fatal("e.HasFired() = true; want false")
	}
	if !e.Fire() {
		t.Fatal("e.Fire() = false; want true")
	}
	if !e.HasFired() {
		t.Fatal("e.HasFired() = false; want true")
	}
}

func TestOnceEvent_Done(t *testing.T) {
	e := NewOnceEvent()
	select {
	case <-e.Done():
		t.Fatal("e.HasFired() = true; want false")
	default:
	}
	if !e.Fire() {
		t.Fatal("e.Fire() = false; want true")
	}
	select {
	case <-e.Done():
	default:
		t.Fatal("e.HasFired() = false; want true")
	}
}

func TestOnceEvent_MultipleFires(t *testing.T) {
	e := NewOnceEvent()
	if e.HasFired() {
		t.Fatal("e.HasFired() = true; want false")
	}
	if !e.Fire() {
		t.Fatal("e.Fire() = false; want true")
	}
	for i := 0; i < 3; i++ {
		if !e.HasFired() {
			t.Fatal("e.HasFired() = false; want true")
		}
		if e.Fire() {
			t.Fatal("e.Fire() = true; want false")
		}
	}
}
